requirements.txt
claude-agent-sdk>=0.1.18
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
google-cloud-firestore>=2.19.0
pydantic-settings>=2.6.0
